refactor(tests): rename bUnit test classes and constants from Team to Club
- git mv Components/TeamLogoTests.cs → ClubLogoTests.cs; rename class + constructor + constant TeamLogoBaseUrl→ClubLogoBaseUrl
- git mv Features/Teams/ → Features/Clubs/ (bUnit); rename classes TeamsIndexTests→ClubsIndexTests, TeamDetailsPageTests→ClubDetailsPageTests, CreateTeamPageTests→CreateClubPageTests, EditTeamPageTests→EditClubPageTests
- Rename private mock handler classes (CreateTeamPageMockHandler→CreateClubPageMockHandler, EditTeamPageMockHandler→EditClubPageMockHandler, NullTeamHttpMessageHandler→NullClubHttpMessageHandler) and their instantiation sites
- Update TeamLogoBaseUrl constant to ClubLogoBaseUrl in all moved bUnit files
- Preserve Features/TeamCompetition/* untouched
